Fixed a few bugs with output and experiment generation.
Hokule'a has problems giving stdout and stderr back when
  they are specified with #PBS -o or #PBS -e.
Letting the scheduler do the default thing works.

cluster_experiment.py wasn't correctly labeling directories
  for aggregated tests. i.e. the same directory would be used for two or more tests.

tensorflow_test.py has been added as a simple device checker to
  make sure that only the expected gpus are visible.

the launcher has been updated to accept a PBS account, number of experiments per job,
  and has some sanity checks around recursive directory deletion. (Deletion bit me)

+++ samples/resnet_cifar_study/resnet_cifar_cluster_experiment_launcher.py
@@ -17,15 +17,21 @@ from cluster_experiment import ClusterExperiment
 def main(FLAGS):
 
     # Clear and remake the log directory.
-    if tf.gfile.Exists(FLAGS.log_dir):
+    if tf.gfile.Exists(FLAGS.log_dir) and FLAGS.delete_log_dir:
 
-        tf.gfile.DeleteRecursively(FLAGS.log_dir)
+        if not FLAGS.delete_without_asking:
+            delete_log_dir = input("Recursively delete directory {} (y/N)?".format(FLAGS.log_dir))
+            
+        if delete_log_dir.lower() == 'y':
+            import pdb
+            pdb.set_trace()
+            tf.gfile.DeleteRecursively(FLAGS.log_dir)
 
     tf.gfile.MakeDirs(FLAGS.log_dir)
 
     # Instantiate an experiment.
     exp = ClusterExperiment()
-
+    
     # Set the number of reps for each config.
     # Set the number of reps for each config.
     exp.set_rep_count(5)
@@ -42,11 +48,12 @@ def main(FLAGS):
     # Launch the experiment.
     exp.launch_experiment(exp_filename=FLAGS.experiment_py_file,
                           log_dir=FLAGS.log_dir,
-                          account='MHPCC96650DE1',
+                          account='MHPCC96650DE1' if not FLAGS.account else FLAGS.account,
                           queue='standard',
                           command_prologue='module load anaconda3/5.0.1 tensorflow',
                           manager='pbs',
-                          shuffle_job_order=True)
+                          shuffle_job_order=True,
+                          experiments_per_job=FLAGS.experiments_per_job)
 
     # Manually note response varaibles.
     response_labels = ['step_num',
@@ -70,6 +77,8 @@ if __name__ == '__main__':
     parser.add_argument('--log_dir', type=str,
                         default='/gpfs/projects/ml/log/resnet_cifar_study/',
                         help='Summaries log directory.')
+    parser.add_argument('--delete_log_dir', default=False, help='Recursively delete the log dir if it exists')
+    parser.add_argument('--delete_without_asking', default=False, help='If deleting the log dir, remove it without asking')
 
     parser.add_argument('--log_filename', type=str,
                         default='resnet_cifar_study.csv',
@@ -83,7 +92,11 @@ if __name__ == '__main__':
                         default='/gpfs/projects/ml/hpc-tensorflow/resnet_cifar_study/resnet_cifar_experiment.py',
                         help='Number of seconds to run before giving up.')
 
+    parser.add_argument('--account', type=str, default=None, help="The PBS Account that this set of experiments should run under.")
+    parser.add_argument('--experiments_per_job', type=int, default=1, help='How many experiments should be batched into a single job')
+    
     # Parse known arguements.
     FLAGS, unparsed = parser.parse_known_args()
 
     main(FLAGS)
+
